export empty Factlabel
